# extensions/issuer-service/membership-attestation-api/src/main/java/org/eclipse/edc/issuerservice/api/MembershipAttestationDto.java
package org.eclipse.edc.issuerservice.api;

import com.fasterxml.jackson.annotation.JsonProperty;

public record MembershipAttestationDto(
        @JsonProperty("id") String id,
        @JsonProperty("name") String name,
        @JsonProperty("holderId") String holderId,
        @JsonProperty("membershipType") String membershipType
) {
}